Moved RestRoutehandler to rest directory.
