+++ src/HeadlessDemo/StartupTask.cs
@@ -2,6 +2,7 @@
 using Devkoes.Restup.WebServer;
 using Windows.ApplicationModel.Background;
 using Devkoes.Restup.WebServer.Http;
+using Devkoes.Restup.WebServer.Rest;
 
 // The Background Application template is documented at http://go.microsoft.com/fwlink/?LinkID=533884&clcid=0x409
